* Added dummy client, so we don't have to worry about null exceptions when sending stuff to creatures. * Added Counter handler. * Fixed Composing (since r143).

+++ src/WorldServer/Network/Sending/Client.Skills.cs
@@ -236,7 +236,7 @@ namespace Aura.World.Network
 		{
 			var p = new MabiPacket(Op.SkillStackSet, creature.Id);
 			p.PutBytes(creature.ActiveSkillStacks, 1);
-			p.PutShort(creature.ActiveSkillId);
+			p.PutShort((ushort)creature.ActiveSkillId);
 
 			client.Send(p);
 		}
